fix: Disable AutoApplyResourcePackFeature (until 1.21/NeoForge)
Forge's Mixin is still SpongePowered Mixin, which is much behind Fabric's fork of mixin. In 1.21, using NeoForge, we won't have this problem, as they use the same Mixin as Fabric.
